refactor(adoc-core): unit 9 — parser/ → infrastructure/parser/
Move all three parser files under infrastructure. Parser files already
used crate::domain::* paths. Update compile.rs and validate.rs test
block to use crate::infrastructure::parser::parse_page.

Refs #27.
