+++ crates/adoc-core/src/infrastructure/mod.rs
@@ -1,4 +1,5 @@
 // see ADR-0009
 pub(crate) mod artifact;
+pub(crate) mod parser;
 pub(crate) mod render;
 pub(crate) mod source;
