+++ crates/adoc-core/src/infrastructure/parser/mod.rs
