+++ crates/adoc-core/src/infrastructure/parser/state.rs
